+++ .github/workflows/clear-and-update-index.yml
@@ -29,7 +29,7 @@ jobs:
           RECORD_MANAGER_DB_URL: ${{ secrets.RECORD_MANAGER_DB_URL }}
 
       - name: Ingest docs
-        run: poetry run python ingest.py
+        run: poetry run python backend/ingest.py
         env:
           OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
           WEAVIATE_URL: ${{ secrets.WEAVIATE_URL }}
